requirements.txt
asgiref==3.8.1
asttokens==2.4.1
cfgv==3.4.0
colorama==0.4.6
crispy-bootstrap5==2024.2
distlib==0.3.8
Django==5.1
django-cors-headers==4.4.0
django-crispy-forms==2.3
django-debug-toolbar==4.4.6
django-extensions==3.2.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
executing==2.0.1
Faker==26.3.0
filelock==3.15.4
icecream==2.1.3
identify==2.6.0
inflection==0.5.1
MarkupSafe==2.1.5
nodeenv==1.9.1
packaging==24.1
pillow==10.4.0
pip==24.2
platformdirs==4.2.2
pre-commit==3.8.0
Pygments==2.18.0
PyJWT==2.9.0
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.2
setuptools==72.1.0
six==1.16.0
sqlparse==0.5.1
uritemplate==4.1.1
virtualenv==20.26.3
Werkzeug==3.0.3
wheel==0.44.0
